Move deprecated simulate_experiment to deprecation.py
