+++ baybe/simulation/core.py
@@ -1,5 +1,4 @@
 """Core simulation and backtesting functionality."""
-import warnings
 from copy import deepcopy
 from dataclasses import dataclass
 from functools import partial
@@ -463,13 +462,3 @@ def simulate_campaign(
         results[cumbest_cols] = cum_fun(results[iterbest_col])
 
     return results
-
-
-def simulate_experiment(*args, **kwargs) -> pd.DataFrame:
-    """A :func:`simulate_campaign` alias for backward compatibility."""  # noqa: D401
-    warnings.warn(
-        "Using 'simulate_experiment' is deprecated and will be removed in a future "
-        "version. Please use the 'simulate_campaign' instead.",
-        DeprecationWarning,
-    )
-    return simulate_campaign(*args, **kwargs)
